refactor: 에러코드명 변경(#3)
