+++ VoiceBook/src/main/java/com/sj/voicebook/member/service/validator/EmailDuplicationValidator.java
@@ -1,11 +1,10 @@
-package com.sj.voicebook.member.service;
+package com.sj.voicebook.member.service.validator;
 
 import com.sj.voicebook.global.exception.BusinessException;
 import com.sj.voicebook.global.exception.ErrorCode;
 import com.sj.voicebook.member.repository.MemberRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
-import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.util.StringUtils;
 
@@ -17,7 +16,7 @@ public class EmailDuplicationValidator {
     @Transactional(readOnly = true)
     public boolean isEmailDuplicated(String email) {
         if(!StringUtils.hasText(email)){
-            throw new BusinessException(ErrorCode.NOT_OR_NULL_FILE_REQUIRED);
+            throw new BusinessException(ErrorCode.NOT_OR_NULL_INPUT_REQUIRED);
         }
         return memberRepository.existsByEmail(email);
     }
